added decent data handling and rendering
